feat: add new types module
