+++ bdk-ffi/src/types.rs
@@ -1,0 +1,148 @@
+use crate::bitcoin::{Address, OutPoint, Script, TxOut};
+
+use bdk::wallet::AddressIndex as BdkAddressIndex;
+use bdk::wallet::AddressInfo as BdkAddressInfo;
+use bdk::wallet::Balance as BdkBalance;
+use bdk::KeychainKind;
+
+use bdk::LocalUtxo as BdkLocalUtxo;
+
+use std::sync::Arc;
+
+/// A output script and an amount of satoshis.
+pub struct ScriptAmount {
+    pub script: Arc<Script>,
+    pub amount: u64,
+}
+
+/// A derived address and the index it was found at.
+pub struct AddressInfo {
+    /// Child index of this address.
+    pub index: u32,
+    /// Address.
+    pub address: Arc<Address>,
+    /// Type of keychain.
+    pub keychain: KeychainKind,
+}
+
+impl From<BdkAddressInfo> for AddressInfo {
+    fn from(address_info: BdkAddressInfo) -> Self {
+        AddressInfo {
+            index: address_info.index,
+            address: Arc::new(address_info.address.into()),
+            keychain: address_info.keychain,
+        }
+    }
+}
+
+/// The address index selection strategy to use to derived an address from the wallet's external
+/// descriptor.
+pub enum AddressIndex {
+    /// Return a new address after incrementing the current descriptor index.
+    New,
+    /// Return the address for the current descriptor index if it has not been used in a received
+    /// transaction. Otherwise return a new address as with AddressIndex::New.
+    /// Use with caution, if the wallet has not yet detected an address has been used it could
+    /// return an already used address. This function is primarily meant for situations where the
+    /// caller is untrusted; for example when deriving donation addresses on-demand for a public
+    /// web page.
+    LastUnused,
+    /// Return the address for a specific descriptor index. Does not change the current descriptor
+    /// index used by `AddressIndex::New` and `AddressIndex::LastUsed`.
+    /// Use with caution, if an index is given that is less than the current descriptor index
+    /// then the returned address may have already been used.
+    Peek { index: u32 },
+}
+
+impl From<AddressIndex> for BdkAddressIndex {
+    fn from(address_index: AddressIndex) -> Self {
+        match address_index {
+            AddressIndex::New => BdkAddressIndex::New,
+            AddressIndex::LastUnused => BdkAddressIndex::LastUnused,
+            AddressIndex::Peek { index } => BdkAddressIndex::Peek(index),
+        }
+    }
+}
+
+impl From<BdkAddressIndex> for AddressIndex {
+    fn from(address_index: BdkAddressIndex) -> Self {
+        match address_index {
+            BdkAddressIndex::New => AddressIndex::New,
+            BdkAddressIndex::LastUnused => AddressIndex::LastUnused,
+            _ => panic!("Mmmm not working"),
+        }
+    }
+}
+
+// TODO 9: Peek is not correctly implemented
+impl From<&AddressIndex> for BdkAddressIndex {
+    fn from(address_index: &AddressIndex) -> Self {
+        match address_index {
+            AddressIndex::New => BdkAddressIndex::New,
+            AddressIndex::LastUnused => BdkAddressIndex::LastUnused,
+            AddressIndex::Peek { index } => BdkAddressIndex::Peek(*index),
+        }
+    }
+}
+
+impl From<&BdkAddressIndex> for AddressIndex {
+    fn from(address_index: &BdkAddressIndex) -> Self {
+        match address_index {
+            BdkAddressIndex::New => AddressIndex::New,
+            BdkAddressIndex::LastUnused => AddressIndex::LastUnused,
+            _ => panic!("Mmmm not working"),
+        }
+    }
+}
+
+pub struct Balance {
+    // All coinbase outputs not yet matured
+    pub immature: u64,
+    /// Unconfirmed UTXOs generated by a wallet tx
+    pub trusted_pending: u64,
+    /// Unconfirmed UTXOs received from an external wallet
+    pub untrusted_pending: u64,
+    /// Confirmed and immediately spendable balance
+    pub confirmed: u64,
+    /// Get sum of trusted_pending and confirmed coins
+    pub trusted_spendable: u64,
+    /// Get the whole balance visible to the wallet
+    pub total: u64,
+}
+
+impl From<BdkBalance> for Balance {
+    fn from(bdk_balance: BdkBalance) -> Self {
+        Balance {
+            immature: bdk_balance.immature,
+            trusted_pending: bdk_balance.trusted_pending,
+            untrusted_pending: bdk_balance.untrusted_pending,
+            confirmed: bdk_balance.confirmed,
+            trusted_spendable: bdk_balance.trusted_spendable(),
+            total: bdk_balance.total(),
+        }
+    }
+}
+
+pub struct LocalUtxo {
+    pub outpoint: OutPoint,
+    pub txout: TxOut,
+    pub keychain: KeychainKind,
+    pub is_spent: bool,
+}
+
+impl From<BdkLocalUtxo> for LocalUtxo {
+    fn from(local_utxo: BdkLocalUtxo) -> Self {
+        LocalUtxo {
+            outpoint: OutPoint {
+                txid: local_utxo.outpoint.txid.to_string(),
+                vout: local_utxo.outpoint.vout,
+            },
+            txout: TxOut {
+                value: local_utxo.txout.value,
+                script_pubkey: Arc::new(Script(local_utxo.txout.script_pubkey)),
+            },
+            keychain: local_utxo.keychain,
+            is_spent: local_utxo.is_spent,
+        }
+    }
+}
